afanasy: fix memory issue processing AFANASY header
previously, sscanf was reading beyond the buffer size as the buffer was
not guaranteed to have a null terminator.

By using a smaller temporary buffer sscanf is considerably faster and
with a guaranteed null terminator there are no problems with parsing the
header.

Replacing sscanf with strtoul should also increase performance and
allows much cleaner error handling.